requirements.txt
django 
pillow
requests
tensorflow